Moved find_resource method to util
